+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove metadata property from JSON before PUTting settings. UICHKOUT-15.
 * Add Patron info. UICHKOUT-5.
 * Remove fixed dependency on plugin-find-users 1.0.0. Refs STRIPES-478, Fixes UICHKOUT-18. 
+* Use PropTypes, not React.PropTypes. Refs STRIPES-427.
 
 ## [1.1.2](https://github.com/folio-org/ui-checkout/tree/v1.1.2) (2017-09-02)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.1.1...v1.1.2)